Improve code coverage to pass SonarCloud validation
