+++ src/main/java/com/justifiedsolutions/justpdf/pdf/contents/AppendToPath.java
@@ -9,16 +9,13 @@ import com.justifiedsolutions.justpdf.pdf.object.PDFReal;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.Objects;
 
 /**
  * Implements the PDF command {@code l} to append a straight line segment to the current path in a content stream.
  *
  * @see "ISO 32000-1:2008, 8.5.2.1"
  */
-public final class AppendToPath implements PathConstructionGraphicsOperator {
-    private final PDFReal x;
-    private final PDFReal y;
+public final class AppendToPath extends LocationOperator implements PathConstructionGraphicsOperator {
 
     /**
      * Appends a new straight line segment to the current path from the previous point to the specified point.
@@ -27,33 +24,14 @@ public final class AppendToPath implements PathConstructionGraphicsOperator {
      * @param y the y value of the next point
      */
     public AppendToPath(PDFReal x, PDFReal y) {
-        this.x = Objects.requireNonNull(x);
-        this.y = Objects.requireNonNull(y);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(x, y);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        AppendToPath startPath = (AppendToPath) o;
-        return x.equals(startPath.x) &&
-                y.equals(startPath.y);
+        super(x, y);
     }
 
     @Override
     public void writeToPDF(OutputStream pdf) throws IOException {
-        x.writeToPDF(pdf);
+        getX().writeToPDF(pdf);
         pdf.write(' ');
-        y.writeToPDF(pdf);
+        getY().writeToPDF(pdf);
         pdf.write(' ');
         pdf.write('l');
         pdf.write('\n');
